+++ packages/sdks/react-sdk/src/types.ts
@@ -93,6 +93,7 @@ export interface IContext {
   session: string;
   isAuthenticated: boolean;
   isSessionLoading: boolean;
+  isOidcLoading: boolean;
   isSessionFetched: boolean;
   projectId: string;
   baseUrl?: string;
@@ -131,6 +132,7 @@ export type DescopeProps = {
   // use to override client context in flow execution
   client?: Record<string, any>;
   styleId?: string;
+  dismissScreenErrorOnInput?: boolean;
   onScreenUpdate?: (
     screenName: string,
     context: Record<string, any>,