+++ packages/sdks/web-js-sdk/src/sdk/index.ts
@@ -1,54 +1,72 @@
-import createCoreSdk from '@descope/core-js-sdk';
+import createCoreSdk, { SdkResponse } from '@descope/core-js-sdk';
 import createWebAuthn from './webauthn';
 import createFedCM from './fedcm';
 import withFlow from './flow';
-import { getSessionToken } from '../enhancers/withPersistTokens/helpers';
+import {
+  getSessionToken,
+  getRefreshToken,
+} from '../enhancers/withPersistTokens/helpers';
+import createOidc from './oidc';
+import { CoreSdk, WebSdkConfig } from '../types';
+import { OIDC_LOGOUT_ERROR_CODE, OIDC_REFRESH_ERROR_CODE } from '../constants';
 
-const createSdk = ({
-  getExternalAccessToken,
-  ...arg
-}: Parameters<typeof createCoreSdk>[0] & {
+const createSdk = (config: WebSdkConfig & {
   getExternalAccessToken?: () => Promise<string>;
 }) => {
-  const coreSdk = createCoreSdk(arg);
+  const coreSdk = createCoreSdk(config);
 
-  console.log('@@@ web-js createSdk with', {
-    getExternalAccessToken,
-  });
+  const oidc = createOidc(coreSdk, config.projectId, config.oidcConfig);
+
   return {
     ...coreSdk,
-    refresh: async (token?: string) => {
+    refresh: async (token?: string): ReturnType<CoreSdk['refresh']> => {
+      if (config.oidcConfig) {
+        try {
+          await oidc.refreshToken(token);
+          return Promise.resolve({ ok: true });
+        } catch (error) {
+          return Promise.resolve({
+            ok: false,
+            error: {
+              errorCode: OIDC_REFRESH_ERROR_CODE,
+              errorDescription: error.toString(),
+            },
+          });
+        }
+      }
       // Descope use this query param to monitor if refresh is made
       // When the user is already logged in in the past or not (We want to optimize that in the future)
       const currentSessionToken = getSessionToken();
-
-      console.log('@@@ calling refresh with', {
-        getExternalAccessToken,
-      });
-      let externalToken = '';
-      if (getExternalAccessToken) {
+      const currentRefreshToken = getRefreshToken();
+      return coreSdk.refresh(token, {
+        dcs: currentSessionToken ? 't' : 'f',
+        dcr: currentRefreshToken ? 't' : 'f',
+      }, await config?.getExternalAccessToken());
+    },
+    // Call the logout function according to the oidcConfig
+    // And return the response in the same format
+    logout: async (token?: string): Promise<SdkResponse<never>> => {
+      if (config.oidcConfig) {
+        // logout is made with id_token_hint
         try {
-          const externalAccessToken = await getExternalAccessToken();
-          console.log('@@@ externalAccessToken', externalAccessToken);
-          if (externalAccessToken) {
-            externalToken = externalAccessToken;
-          }
-        } catch (e) {
-          console.error('Failed to get external access token', e);
+          await oidc.logout({ id_token_hint: token });
+          return Promise.resolve({ ok: true });
+        } catch (error) {
+          return Promise.resolve({
+            ok: false,
+            error: {
+              errorCode: OIDC_LOGOUT_ERROR_CODE,
+              errorDescription: error.toString(),
+            },
+          });
         }
       }
-
-      return coreSdk.refresh(
-        token,
-        {
-          dcs: currentSessionToken ? 't' : 'f',
-        },
-        externalToken,
-      );
+      return coreSdk.logout(token);
     },
     flow: withFlow(coreSdk),
     webauthn: createWebAuthn(coreSdk),
-    fedcm: createFedCM(coreSdk, arg.projectId),
+    fedcm: createFedCM(coreSdk, config.projectId),
+    oidc,
   };
 };