+++ packages/widgets/access-key-management-widget/CHANGELOG.md
@@ -2,6 +2,53 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.11](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.10...access-key-management-widget-0.4.11) (2025-03-27)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.29.0`
+## [0.4.10](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.9...access-key-management-widget-0.4.10) (2025-03-26)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.11.3`
+
+### Bug Fixes
+
+* Add formatted time to widget mocks RELEASE ([#1066](https://github.com/descope/descope-js/issues/1066)) ([739ac9e](https://github.com/descope/descope-js/commit/739ac9e821a3e2ead082782c91922817ab38c567))
+
+## [0.4.9](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.8...access-key-management-widget-0.4.9) (2025-03-21)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.28.0`
+## [0.4.8](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.7...access-key-management-widget-0.4.8) (2025-03-19)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.2`
+## [0.4.7](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.6...access-key-management-widget-0.4.7) (2025-03-17)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.1`
+## [0.4.6](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.5...access-key-management-widget-0.4.6) (2025-03-13)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.0`
+## [0.4.5](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.4...access-key-management-widget-0.4.5) (2025-03-11)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.3.0`
+* `sdk-mixins` updated to version `0.11.2`
+* `sdk-component-drivers` updated to version `0.2.52`
+
+### Bug Fixes
+
+* issue 9615 ([#1042](https://github.com/descope/descope-js/issues/1042)) ([d1eb29a](https://github.com/descope/descope-js/commit/d1eb29a3a96a5c0f4ac9859104cf8d3bf5d06d6d))
+
 ## [0.4.4](https://github.com/descope/descope-js/compare/access-key-management-widget-0.4.3...access-key-management-widget-0.4.4) (2025-03-06)
 
 ### Dependency Updates