+++ packages/widgets/user-management-widget/CHANGELOG.md
@@ -2,6 +2,48 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.7.11](https://github.com/descope/descope-js/compare/user-management-widget-0.7.10...user-management-widget-0.7.11) (2025-03-27)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.29.0`
+## [0.7.10](https://github.com/descope/descope-js/compare/user-management-widget-0.7.9...user-management-widget-0.7.10) (2025-03-26)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.11.3`
+
+### Bug Fixes
+
+* Add formatted time to widget mocks RELEASE ([#1066](https://github.com/descope/descope-js/issues/1066)) ([739ac9e](https://github.com/descope/descope-js/commit/739ac9e821a3e2ead082782c91922817ab38c567))
+
+## [0.7.9](https://github.com/descope/descope-js/compare/user-management-widget-0.7.8...user-management-widget-0.7.9) (2025-03-21)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.28.0`
+## [0.7.8](https://github.com/descope/descope-js/compare/user-management-widget-0.7.7...user-management-widget-0.7.8) (2025-03-19)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.2`
+## [0.7.7](https://github.com/descope/descope-js/compare/user-management-widget-0.7.6...user-management-widget-0.7.7) (2025-03-17)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.1`
+## [0.7.6](https://github.com/descope/descope-js/compare/user-management-widget-0.7.5...user-management-widget-0.7.6) (2025-03-13)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.27.0`
+## [0.7.5](https://github.com/descope/descope-js/compare/user-management-widget-0.7.4...user-management-widget-0.7.5) (2025-03-11)
+
+### Dependency Updates
+
+* `sdk-helpers` updated to version `0.3.0`
+* `sdk-mixins` updated to version `0.11.2`
+* `sdk-component-drivers` updated to version `0.2.52`
 ## [0.7.4](https://github.com/descope/descope-js/compare/user-management-widget-0.7.3...user-management-widget-0.7.4) (2025-03-06)
 
 ### Dependency Updates